Committing DL
