+++ Deep Learning/Tester.csv
@@ -1,0 +1,20 @@
+Sentence
+"A movie with a really interesting idea and with menacing aliens as enemies. It was mostly fun to watch. Although, it could be a bit messy and rushed sometimes, and had a certain exaggeration at parts."
+"A brilliant film with a brilliant idea and a terrifying enemy, that might be called evil itself. Sadly, the film was executed very poorly. I disliked most of the characterization and the movie became more and more of a mess as it went towards the end. Giving it an average rating though, because of the story arch."
+"This movie will give you headache in your liver and heartattack in your brain. Forget the logic, i want sefloon bhai to go back in time and slap Newton for discovering gravity, 'cause here he nothing but rock bed with it. And my god bhai's rabid fan base. They would watch such films 4-5 times and but won't donate that money to charity."
+"It's utter ridiculousness that this film is so hated. I don't know why it's hated. Perhaps it's comparing to previous films or expectations, but that should have nothing to do with how good this film itself is. I have learned that from now on never care about what the general opinion of a movie is, because the general movie judger is a ridiculous, nostalgic and emotional fool that thinks a movie is either super bad or super good, especially if the movie shows different things than what the watcher had in mind. This is also the last review on this particular list, and I reviewed this movie mostly to get some chance to get these extra words concerning the movie out there."
+"Acting, yes acting. It's quite odd to see capable actors like Ranbir and Alia act so loudly and throw each and every dialogue out loud. Same for Mouni Roy who is screaming most of the time. There could a be a bit more toned down acting."
+"The Shadow Within fails to live up to its intriguing premise. Despite commendable visual effects, the convoluted plot and lackluster character development make it a forgettable experience. Viewers might find themselves lost in the maze of subplots, ultimately leaving them unsatisfied."
+"Echoes of Eternity transports audiences to a breathtaking realm of cinematic wonder. With its mesmerizing cinematography and emotionally resonant storyline, the film is a triumph in fantasy storytelling. Each frame is a masterpiece, and the performances are nothing short of extraordinary. A true gem for lovers of epic adventures."
+"The Silent Hill misses the mark on every level. Its attempt to revive a classic franchise falls flat, with wooden acting and a predictable plotline. The scares are telegraphed, and the atmosphere feels forced rather than immersive. A disappointing entry that fails to capture the essence of its source material."
+"Sunset Serenade is a heartwarming tale that resonates long after the credits roll. Set against the backdrop of a quaint coastal town, the film weaves together themes of love, loss, and redemption with grace and sincerity. The stellar performances and beautiful soundtrack elevate this cinematic gem to must-see status."
+"This film is a haunting melody that stays with you long after the credits roll. While its cinematography and score are mesmerizing, the plot feels disjointed at times, leaving viewers grasping for coherence. Despite its flaws, 'The Moonlit Sonata' is a unique cinematic experience that deserves appreciation for its artistic vision."""
+"A promising premise drowned in clichs and predictable twists. 'Lost in the Echoes' struggles to find its identity amidst a sea of forgettable performances and uninspired storytelling. Despite decent production values, this film fails to leave a lasting impression, echoing the mediocrity of countless Hollywood dramas."
+"I couldn't wait for this movie to end. The pacing was painfully slow, the characters were one-dimensional, and the storyline was predictable. I found myself checking my watch constantly, just hoping it would be over soon."
+"I absolutely loved this movie! It had everything I could ask for - action, romance, humor, and heart. The chemistry between the leads was electric, and the plot twists kept me guessing until the very end. I'll definitely be watching it again."
+"What a masterpiece! The acting was superb, the storyline was gripping, and the special effects were mind-blowing. I was on the edge of my seat throughout the entire movie. Definitely a must-see!"
+"Honestly, I was quite disappointed with this movie. The plot was convoluted, the acting felt forced, and the dialogue was cheesy. I kept waiting for it to get better, but it never did. Save your money and watch something else."
+"This movie exceeded all my expectations! The cinematography was breathtaking, the soundtrack was captivating, and the characters were so well-developed. I laughed, I cried, and I left the theater feeling inspired. A true cinematic gem."
+"I wish I could unsee this movie. The plot was nonsensical, the acting was wooden, and the special effects were laughably bad. It felt like a waste of time and money. I wouldn't recommend it to anyone."
+"This movie was a breath of fresh air! The storyline was original, the characters were relatable, and the humor was spot-on. I found myself smiling from ear to ear throughout the entire film. It's rare to find a movie that's both entertaining and thought-provoking, but this one nailed it!"
+"I have to say, I was extremely underwhelmed by this movie. The marketing hyped it up to be the next big thing, but it fell completely flat. The dialogue was cringe-worthy, and the acting was subpar at best. I left the theater feeling disappointed and wishing I had spent my time elsewhere."
